# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Trita API
  version: "2026-08-23"
  description: |
    Trita platform belső API-leírója — kézzel karbantartott, a route-fájlok
    Zod-sémáiból származtatva. Auth: Clerk session cookie (böngésző-hívások),
    kivéve a jelölt publikus, token-alapú végpontokat.

    Csoportok (tags): Public · Assessment · Observer · Profile · Notifications
    · Feedback · Team · TeamRoles/Trust/PsychSafety · Org · Manager/Hiring
    · Admin · Inquiry · System.

    Generálás: a projekt sima Next.js route-handlereket használ (nincs
    dekorátor-réteg), ezért teljesen automatikus generátor nincs bekötve.
    Ajánlott irány: zod-openapi (a meglévő Zod-sémák annotálása) vagy
    next-rest-framework — lásd docs/security/security-audit-2026-07-22.md.
servers:
  - url: https://trita.io
  - url: http://localhost:3000
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk session cookie — a böngésző automatikusan küldi.
    cronBearer:
      type: http
      scheme: bearer
      description: CRON_SECRET Bearer tokenként (Vercel cron).
    svixSignature:
      type: apiKey
      in: header
      name: svix-signature
      description: Svix webhook aláírás (svix-id + svix-timestamp + svix-signature).
...
security:
  - clerkSession: []
...
paths:
...
  /api/team/{id}/invite:
    post:
      tags: [Team]
      summary: Email-meghívó csapatba (TeamPendingInvite + join-link)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses: { "200": { description: "{ signUpUrl }" } }
...